# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manage OAuth-connected Gmail and Microsoft mailboxes, proxies, and synced messages.",
        "title": "DeezMails API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/accounts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List mailbox accounts",
                "parameters": [
                    {
                        "enum": [
                            "gmail",
                            "microsoft"
                        ],
                        "type": "string",
                        "description": "Filter provider",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Account"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create mailbox account metadata",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AccountInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    }
                }
            }
        },
        "/api/accounts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Update an account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account update",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AccountUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "accounts"
                ],
                "summary": "Delete account and synced messages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Account": {
            "type": "object",
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "folder": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incomingHost": {
                    "type": "string"
                },
                "incomingPort": {
                    "type": "integer"
                },
                "lastCheckedAt": {
                    "type": "string"
                },
                "lastError": {
                    "type": "string"
                },
                "lastErrorAction": {
                    "$ref": "#/definitions/models.MailErrorAction"
                },
                "lastErrorCode": {
                    "$ref": "#/definitions/models.MailErrorCode"
                },
                "lastSyncedAt": {
                    "type": "string"
                },
                "nextSyncAt": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "proxy": {
                    "$ref": "#/definitions/models.Proxy"
                },
                "proxyId": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "syncEnabled": {
                    "type": "boolean"
                },
                "tlsMode": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AccountInput": {
            "type": "object",
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "folder": {
                    "type": "string"
                },
                "incomingHost": {
                    "type": "string"
                },
                "incomingPort": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "proxyId": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "tlsMode": {
                    "type": "string"
                }
            }
        },
        "models.AccountUpdateInput": {
            "type": "object",
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "folder": {
                    "type": "string"
                },
                "incomingHost": {
                    "type": "string"
                },
                "incomingPort": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "proxyId": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "syncEnabled": {
                    "type": "boolean"
                },
                "tlsMode": {
                    "type": "string"
                }
            }
        },
...
        "models.MailErrorAction": {
            "type": "string",
            "enum": [
                "none",
                "retry",
                "reconnect",
                "edit_account",
                "contact_admin"
            ],
            "x-enum-varnames": [
                "MailErrorActionNone",
                "MailErrorActionRetry",
                "MailErrorActionReconnect",
                "MailErrorActionEditAccount",
                "MailErrorActionContactAdmin"
            ]
        },
        "models.MailErrorCode": {
            "type": "string",
            "enum": [
                "authentication_failed",
                "permission_denied",
                "organization_policy",
                "quota_exceeded",
                "rate_limited",
                "provider_unavailable",
                "not_found",
                "invalid_configuration",
                "tls_failed",
                "network_failed",
                "protocol_failed",
                "response_too_large",
                "cancelled",
                "internal_error"
            ],
            "x-enum-varnames": [
                "MailErrorAuthentication",
                "MailErrorPermission",
                "MailErrorPolicy",
                "MailErrorQuota",
                "MailErrorRateLimit",
                "MailErrorUnavailable",
                "MailErrorNotFound",
                "MailErrorConfiguration",
                "MailErrorTLS",
                "MailErrorNetwork",
                "MailErrorProtocol",
                "MailErrorResponseTooLarge",
                "MailErrorCancelled",
                "MailErrorInternal"
            ]
        },
        "models.Proxy": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}